requirements.txt
openai>=1.0.0
python-dotenv
SpeechRecognition
PyAudio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dateparser
pytz